==> Automated/Code/AutomatedSteps/Step0.ipynb <==
# import netCDF4
import pandas as pd
import datetime as dt
import numpy as np
import csv
import calendar

import os,sys,inspect
current_dir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
parent_dir = os.path.dirname(current_dir)
sys.path.insert(0, parent_dir) 
from environmentVariables import environment_variables


userTemp = os.getlogin()

if userTemp == 'Eshta':
    user = 'eshta'
user_key=environment_variables[user]
pd.set_option('display.precision', 12)

Userdata=pd.read_csv('AutomatedSteps/userInput.csv', header=[0])
SummaryDF = pd.DataFrame(columns=['Variable'])
directory_list = list()

for root, dirs, files in os.walk(user_key['path_in'], topdown=False):
    for name in dirs:
        directory_list.append(name)

print(directory_list)

for col in Userdata.columns:
    #if the column name exists in the data variables then add to list
    
    if col == 'T2M':
        col = 'tas'
    elif col == 'PRECTOTCORR' :
        col = 'pr'
    elif col == 'PS' :
        col = 'ps'
    elif col == 'WS10M' :
        col = 'sfcWind'
    elif col == 'WS10M_MAX' :
        col = 'sfcWindmax'
    elif col == 'T2M_MAX' :
        col = 'tasmax'
    elif col == 'T2M_MIN' :
        col = 'tasmin'
    elif col == 'ALLSKY_SFC_LW_DWN':
        col = 'rlds'
    #col = col.lower()
    print(col)
    if col in directory_list:
        
        SummaryDF = SummaryDF.append({'Variable':col}, ignore_index=True)

print(SummaryDF)



#logic to create an output for data 

#create counts 

outputDF = pd.DataFrame(columns=['Column'])
NA = []
blanks = []
 
for col in Userdata.columns:
    #use each column
   
    tempNA = 0
    tempBlanks = 0
    outputDF = outputDF.append({'Column': col}, ignore_index=True)
    for index, row in Userdata.iterrows():
        #look at each item in each column
        if Userdata.loc[index,col] == -999 :
            tempNA = tempNA + 1
        
        if Userdata.loc[index,col] == '' :
            tempBlanks = tempBlanks + 1
            
    print(tempNA)
    NA.append(tempNA)
    blanks.append(tempBlanks)

print(NA)
print(outputDF)
print(blanks)
outputDF['Nulls'] = NA
outputDF['Blanks'] = blanks 
    





SummaryDF.to_csv('AutomatedSteps/variableList.csv', index=None)
outputDF.to_csv('AutomatedSteps/outputFeedback.csv', index=None)
