==> Automated/Code/AutomatedSteps/step2.combine_years_into_one_var_file.ipynb <==
# import netCDF4
import pandas as pd
import datetime as dt
import numpy as np
import csv
import calendar

import os,sys,inspect
current_dir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
parent_dir = os.path.dirname(current_dir)
sys.path.insert(0, parent_dir) 
from environmentVariables import environment_variables

# INPUT

# read files

#this will be based on variable in the user input


selections=pd.read_csv('AutomatedSteps/optionSelection.csv', header=[0])


for index, row in selections.iterrows():
    if selections.at[index,'option'] == 'variable':
        variable = selections.at[index,'selection']

data_one=pd.read_csv('AutomatedSteps/'+ variable + '_1986-1990.csv')
data_two=pd.read_csv('AutomatedSteps/'+ variable + '_1991-1995.csv')
data_three=pd.read_csv('AutomatedSteps/'+ variable + '_1996-2000.csv')
data_four=pd.read_csv('AutomatedSteps/'+ variable + '_2001-2005.csv')


# MERGE

frames = [data_one, data_two, data_three, data_four]
result = pd.concat(frames)

#frames

result.to_csv('AutomatedSteps/'+ variable + '.csv', index=None)

#clean up and delete those files
os.remove('AutomatedSteps/'+ variable + '_1986-1990.csv')
os.remove('AutomatedSteps/'+ variable + '_1991-1995.csv')
os.remove('AutomatedSteps/'+ variable + '_1996-2000.csv')
os.remove('AutomatedSteps/'+ variable + '_2001-2005.csv')
