==> Automated/Code/AutomatedSteps/step3.Unit_Conversion.ipynb <==
# import netCDF4
import pandas as pd
import datetime as dt
import numpy as np
import csv
import calendar

import os,sys,inspect
current_dir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
parent_dir = os.path.dirname(current_dir)
sys.path.insert(0, parent_dir) 
from environmentVariables import environment_variables

# read files
selections=pd.read_csv('AutomatedSteps/optionSelection.csv', header=[0])
for index, row in selections.iterrows():
    if selections.at[index,'option'] == 'variable':
        variable = selections.at[index,'selection']
        
        
data=pd.read_csv('AutomatedSteps/'+variable +'.csv')
observedData=pd.read_csv('AutomatedSteps/userInput.csv')

# conversion equation based on variable name
variableName = variable

#print(data)


# loop through the rows creating a new converted column    
row_iterator = data.iterrows()   
for i, row in row_iterator:
    currVar = (data.at[i,'Variable'])
    if (currVar != "Variable" ):
        if (variableName == "pr"):
            #1 kg/m2/s = 86400 mm/day
            data.at[i, 'Variable_converted'] = (float(currVar) *  (86400))
        elif (variableName == "ps"):
            #1 pa	0.001 kP
             data.at[i, 'Variable_converted'] = (float(currVar) *  (0.001))
        elif (variableName == "rlds"):
            data.at[i, 'Variable_converted'] = (float(currVar))
        elif (variableName == "sfcWind"):
            #1 m/s	3.6km/h
            data.at[i, 'Variable_converted'] = (float(currVar) *  (3.6))
        elif (variableName == "sfcWindmax"):
            #1 m/s	3.6km/h
            data.at[i, 'Variable_converted'] = (float(currVar) *  (3.6))
        elif (variableName == "tas"):
            # 1K = -273 degrees celsius
            data.at[i, 'Variable_converted'] = (float(currVar) - (273.15))
        elif (variableName == "tasmax"):
            # 1K = -273 degrees celsius
            data.at[i, 'Variable_converted'] = (float(currVar) - (273.15))
        elif (variableName == "tasmin"):
            # 1K = -273 degrees celsius
            data.at[i, 'Variable_converted'] = (float(currVar) - (273.15))
                                  
        
#print(data)

row_iterator2 = observedData.iterrows()   
for i, row in row_iterator2:
    if (variableName == "sfcWind"):
        currVar = (observedData.at[i,'WS10M'])
        #1 m/s	3.6km/h
        observedData.at[i, 'WS10M'] = (float(currVar) *  (3.6))
    elif (variableName == "sfcWindmax"):
        currVar = (observedData.at[i,'WS10M_MAX'])
        #1 m/s	3.6km/h
        observedData.at[i, 'WS10M_MAX'] = (float(currVar) *  (3.6))
                        
#print(observedData)

# delete unnecessary columns
del data['Variable']

# rename column
data = data.rename(columns={'Variable_converted': 'Variable'})

data

data.to_csv('AutomatedSteps/'+variable +'_converted.csv', index=None)
observedData.to_csv('AutomatedSteps/userInput.csv', index=None)
