==> Automated/Code/AutomatedSteps/step4.mapping_model_to_observed.ipynb <==
import pandas as pd
import datetime as dt
import numpy as np
import csv
import calendar

import dash
import dash_core_components as dcc
import dash_html_components as html

import os,sys,inspect
from dash.dependencies import Input, Output

current_dir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
parent_dir = os.path.dirname(current_dir)
sys.path.insert(0, parent_dir) 
from environmentVariables import environment_variables

#use inputs to select file for climate model
selections=pd.read_csv('AutomatedSteps/optionSelection.csv', header=[0])
for index, row in selections.iterrows():
    if selections.at[index,'option'] == 'variable':
        variable = selections.at[index,'selection']
        

data=pd.read_csv('AutomatedSteps/'+variable+'_converted.csv')



# read observational climate csv provided by user

observeData=pd.read_csv('AutomatedSteps/userInput.csv')


data['Date'] =  pd.to_datetime(data['Date'], format='%Y-%m-%d')

# read column based on the variable 

 
if variable == 'tas':
            #then it is TEMP in the data, find the column
    column = 'T2M'
    
elif variable == 'pr' :
    column = 'PRECTOTCORR'
            
elif variable == 'ps' :
    column = 'PS'
            
elif variable == 'sfcWind' :
    column = 'WS10M'
            
elif variable == 'sfcWindmax' :
    column = 'WS10M_MAX'
            
elif variable == 'tasmax' :
    column = 'T2M_MAX'
            
elif variable == 'tasmin' :
    column = 'T2M_MIN'

elif variable == 'rlds':
    column = 'ALLSKY_SFC_LW_DWN'
          

row_iterator = observeData.iterrows()
counter =0

dates = []
observedData = []
modelData = []

#the for loop will loop through the observed data file (weatherstats)
for i, row in row_iterator:
        if counter != len(observeData):
            dtTemp = str(observeData.at[i,'YEAR']) + "-" + str(observeData.at[i,'MO']) + "-" + str(observeData.at[i,'DY']) 
            if(str(observeData.at[i,'MO']) == '2' and str(observeData.at[i,'DY']) == '29'):
                print("skipped : " + dtTemp)
            else:
                dates.append(dtTemp)
                observedData.append(observeData.at[i,column])
                counter += 1

row_ittr2 = data.iterrows()
counter =0
#the for loop will loop through the model data file 
for i, row in row_ittr2:
         if counter != len(data):
            modelData.append(data.at[i,'Variable'])
            counter += 1


            

#now we have all the variables in array now we create a dataframe
data_all = pd.DataFrame()
data_all.insert(0, 'Date', dates)
data_all.insert(1, 'Observed'+variable, observedData)
data_all.insert(2, 'Model'+variable, modelData)
        
#print(data_all)

data_all.to_csv('AutomatedSteps/'+variable+'_comparison.csv', index=None)
